// todo/kesq2/src/model.d.ts
export type Future<T> = Generator<void, T, void>;
declare class FutureContext {
	#private;
	constructor(coro: Generator);
	wakeup(): void;
	throw(e: Error): void;
}
interface Storage$1 {
	withWTxn<T>(fx: FutureContext, fn: (txn: WTxn) => Future<T>): Future<T>;
	withRTxn<T>(fx: FutureContext, fn: (txn: RTxn) => Future<T>): Future<T>;
}
export type StorageValue = {
	value: unknown;
} | {
	err: Error;
};
export type StorageDone = {
	value: true;
} | {
	err: Error;
};
export interface WTxn {
	get(key: string, cb: (result: StorageValue) => void): void;
	set(key: string, value: unknown, cb: (result: StorageDone) => void): void;
	del(key: string, cb: (result: StorageDone) => void): void;
}
export interface RTxn {
	get(key: string, cb: (result: StorageValue) => void): void;
	set(key: string, value: unknown, cb: (result: StorageDone) => void): void;
	del(key: string, cb: (result: StorageDone) => void): void;
}
export declare class InMemStorage {
	#private;
	constructor(data?: Record<string, unknown>);
	withWTxn<T>(_fx: FutureContext, fn: (txn: WTxn) => Future<T>): Future<T>;
	withRTxn<T>(_fx: FutureContext, fn: (txn: RTxn) => Future<T>): Future<T>;
}
export type ReducerQuestion = {
	old?: Record<string, true>;
	get?: Record<string, true>;
	set?: Record<string, unknown>;
	del?: Record<string, true>;
};
export type ReducerAnswer = {
	old: Record<string, StorageValue>;
	get: Record<string, StorageValue>;
	set: Record<string, StorageDone>;
	del: Record<string, StorageDone>;
};
export type Reducer<T> = Generator<ReducerQuestion, T, ReducerAnswer>;
export interface Query<T> {
	latest: T | undefined;
	awaitResult(): QueryGenerator<T>;
	subscribe(callback: (val: T) => void): () => void;
	start(): void;
	close(): void;
}
export type QueryQuestion = {
	store?: Record<string, true>;
	query?: Record<string, true>;
};
export type QueryAnswer = {
	store: Record<string, StorageValue>;
	query: Record<string, [
		unknown,
		boolean
	]>;
};
export type QueryGenerator<T> = Generator<QueryQuestion, T, QueryAnswer>;
export type QueryFunction<QX, T> = (qx: QX, prev: T | undefined, prevIsValid: boolean) => QueryGenerator<T>;
type Event$1<T> = {
	id: string;
	data: T;
};
export type RealEvent<T> = Event$1<T> & {
	position: number;
};
declare class Framework<QX, RX, E, C> {
	#private;
	constructor(qx: QX, rx: RX, storage: Storage$1 | null, callbacks: {
		decodeEvent: (raw: any) => E;
		decodeCommand: (raw: any) => C;
		migrate?: (rx: RX) => Reducer<void>;
		reducer: (rx: RX, events: E[]) => Reducer<void | any[]>;
		forecaster?: (commands: C) => E[];
		onCommands?: (commands: any[]) => void;
	});
	reconnect(cb: (result: {
		checkpoint: number | undefined;
		commands: Event$1<any>[];
	}) => void): void;
	recvEvents(raw: RealEvent<any>[]): void;
	fellBehind(): void;
	caughtUp(): void;
	sendCommands(commands: C[]): void;
	markSent(...id: string[]): void;
	newQuery<T>(fn: QueryFunction<QX, T>, manualStart?: boolean): Query<T>;
	simulate<T>(fn: (rx: RX, decodedEvents: E[]) => Reducer<T>, cb: (result: T) => void, undecodedEvents?: Event$1<any>[]): void;
}
export type List = {
	id: string;
	name: string;
	items: string[];
	archived: boolean;
};
export type Item = {
	id: string;
	text: string;
	done: boolean;
	archived: boolean;
};
export type NewList = {
	type: "new-list";
	id: string;
	name: string;
};
export type RenameList = {
	type: "rename-list";
	id: string;
	name: string;
};
export type ArchiveList = {
	type: "archive-list";
	id: string;
};
export type NewItem = {
	type: "new-item";
	id: string;
	list: string;
	text: string;
};
export type EditItem = {
	type: "edit-item";
	id: string;
	text: string;
};
export type MarkItem = {
	type: "mark-item";
	id: string;
	done: boolean;
};
export type ArchiveItem = {
	type: "archive-item";
	id: string;
};
export type TodoEvents = EditItem | MarkItem | ArchiveList | NewList | NewItem | ArchiveItem | RenameList;
export declare function DecodeTodoEvents(val: any): TodoEvents;
declare const TodoQueryContext: {
	get: {
		all_lists: () => QueryGenerator<string[]>;
		item: (item_id: string) => QueryGenerator<Item>;
		list: (list_id: string) => QueryGenerator<List>;
	};
};
export type TodoQX = typeof TodoQueryContext;
declare const TodoReducerContext: {
	old: {
		all_lists: () => Reducer<string[]>;
		item: (item_id: string) => Reducer<Item>;
		list: (list_id: string) => Reducer<List>;
	};
	get: {
		all_lists: () => Reducer<string[]>;
		item: (item_id: string) => Reducer<Item>;
		list: (list_id: string) => Reducer<List>;
	};
	set: {
		all_lists: (value: string[]) => Reducer<void>;
		item: (item_id: string, value: Item) => Reducer<void>;
		list: (list_id: string, value: List) => Reducer<void>;
	};
	del: {
		item: (item_id: string) => Reducer<void>;
		list: (list_id: string) => Reducer<void>;
	};
	update: {
		all_lists: <R>(fn: (value: string[]) => R) => Reducer<R>;
		item: <R>(item_id: string, fn: (value: Item) => R) => Reducer<R>;
		list: <R>(list_id: string, fn: (value: List) => R) => Reducer<R>;
	};
};
export type TodoRX = typeof TodoReducerContext;
export declare class TodoFramework extends Framework<TodoQX, TodoRX, TodoEvents, TodoEvents> {
	constructor(storage: Storage$1, callbacks: {
		migrate?: (rx: TodoRX) => Reducer<void>;
		reducer: (rx: TodoRX, events: TodoEvents[]) => Reducer<void | any[]>;
		forecaster?: (commands: TodoEvents) => TodoEvents[];
		onCommands?: (commands: Event$1<any>[]) => void;
	}, qx?: any);
}